+++ spec/unit/installer/pod_source_installer_spec.rb
@@ -132,23 +132,6 @@ module Pod
       end
 
       #--------------------------------------#
-
-      describe 'Locking' do
-        it 'locks the source files for each Pod' do
-          File.expects(:chmod).at_least_once
-          @installer.install!
-          @installer.lock_files!
-        end
-
-        it "doesn't lock local pods" do
-          @installer.stubs(:local?).returns(true)
-          File.expects(:chmod).never
-          @installer.install!
-          @installer.lock_files!
-        end
-      end
-
-      #--------------------------------------#
     end
 
     #-------------------------------------------------------------------------#